hdf5 builder (pass 1)
