+++ cdm/core/src/main/java/ucar/nc2/NetcdfFiles.java
@@ -59,7 +59,7 @@ public class NetcdfFiles {
 
     // Register iosp's that are part of cdm-core. This ensures that they are tried first.
     try {
-      registerIOProvider("ucar.nc2.iosp.hdf5.H5iosp");
+      registerIOProvider("ucar.nc2.internal.iosp.hdf5.H5iospNew");
     } catch (Throwable e) {
       if (loadWarnings)
         log.info("Cant load class H5iosp", e);
